Trait-ify custom fields
